Refactor: Move classes to new cs files
