+++ 1.Simplicity/Qaf/Solutions/DesignPattern/Imp/IFlatTaxCalculator.cs
@@ -1,0 +1,6 @@
+namespace Qaf.Solutions.DesignPattern.Imp;
+
+public interface IFlatTaxCalculator
+{
+    decimal Calculate();
+}
