+++ 1.Simplicity/Qaf/Solutions/DesignPattern/Imp/IRelativeTaxCalculator.cs
@@ -1,0 +1,6 @@
+namespace Qaf.Solutions.DesignPattern.Imp;
+
+public interface IRelativeTaxCalculator
+{
+    decimal Calculate(decimal currentTax);
+}
