+++ 1.Simplicity/Qaf/Solutions/DesignPattern/Imp/ITaxCalculator.cs
@@ -1,0 +1,6 @@
+namespace Qaf.Solutions.DesignPattern.Imp;
+
+public interface ITaxCalculator
+{
+    decimal Calculate();
+}
